+++ finject_flutter/lib/finject_flutter.dart
@@ -1,4 +1,7 @@
 library finject_flutter;
 
+export 'src/finject_host.dart';
+export 'src/flat_inject_host.dart';
 export 'src/inject_host.dart';
 export 'src/injection_provider.dart';
+export 'src/scope_host.dart';