+++ finject_flutter/lib/src/flat_inject_host.dart
@@ -3,39 +3,29 @@ import 'package:flutter/material.dart';
 
 import 'injection_provider.dart';
 
-class InjectHost extends StatelessWidget {
+class JustInjectHost extends StatelessWidget {
   final Widget child;
-  _InjectionProviderImpl _provider;
+  JustInjectionProviderImpl _provider;
 
   @protected
-  InjectHost({this.child}) {
-    _provider = _InjectionProviderImpl();
+  JustInjectHost({this.child}) {
+    _provider = JustInjectionProviderImpl();
   }
 
   @override
   Widget build(BuildContext context) {
-    _provider.context = context;
     _provider.inject(child);
     return child;
   }
 }
 
-class _InjectionProviderImpl extends AbstractInjectionProvider {
-  BuildContext context;
-
-  _InjectionProviderImpl();
+class JustInjectionProviderImpl extends AbstractInjectionProvider {
+  JustInjectionProviderImpl();
 
   @override
   T get<T>([String name]) {
     T value;
     Qualifier qualifier = QualifierFactory.create(T, name);
-
-    FoundInjection foundInjection = findParrent(context as BuildContext);
-    InjectionProvider parentInjector = foundInjection.provider;
-    if (parentInjector != null) {
-      value = parentInjector.get(name);
-      return value;
-    }
 
     Factory factory = rootDependencyResolver["factory"][qualifier];
     if (factory != null) {
@@ -47,13 +37,6 @@ class _InjectionProviderImpl extends AbstractInjectionProvider {
 
   inject(Object target, [String name]) {
     Qualifier qualifier = QualifierFactory.create(target.runtimeType, name);
-
-    FoundInjection foundInjection = findParrent(context as BuildContext);
-    InjectionProvider parentInjector = foundInjection.provider;
-    if (parentInjector != null) {
-      parentInjector.inject(target, name);
-      return;
-    }
 
     Injector injector = rootDependencyResolver["injector"][qualifier];
     if (injector != null) {