+++ finject_flutter/lib/src/injection_provider.dart
@@ -2,79 +2,32 @@
 import 'package:finject/finject.dart';
 import 'package:flutter/material.dart';
 
-import 'inject_host.dart';
+import 'scope_host.dart';
 
-class InjectionProviderImpl extends InjectionProvider {
+abstract class AbstractInjectionProvider extends InjectionProvider {
   BuildContext context;
-  Scope scope;
-
-  InjectionProviderImpl(this.scope);
-
-  @override
-  T get<T>([String name]) {
-    T value;
-
-    Qualifier qualifier = QualifierFactory.create(T, name);
-
-    if (scope != null &&
-        scope.factories[qualifier] != null &&
-        scope.injectors[qualifier] != null) {
-      value = scope.factories[qualifier].create(this);
-      scope.injectors[qualifier].inject(value, this);
-      return value;
-    }
-
-    FoundInjection foundInjection = findParrent(context as BuildContext);
-    InjectionProvider parentInjector = foundInjection.provider;
-    if (parentInjector != null) {
-      value = parentInjector.get(name);
-      return value;
-    }
-
-    Factory factory = rootDependencyResolver["factory"][qualifier];
-    if (factory != null) {
-      value = factory.create(this);
-      rootDependencyResolver["injector"][qualifier].inject(value, this);
-      return value;
-    }
-  }
-
-  inject(Object target, [String name]) {
-    Qualifier qualifier = QualifierFactory.create(target.runtimeType, name);
-
-    if (scope != null && scope.injectors[qualifier] != null) {
-      scope.injectors[qualifier].inject(target, this);
-      return;
-    }
-
-    FoundInjection foundInjection = findParrent(context as BuildContext);
-    InjectionProvider parentInjector = foundInjection.provider;
-    if (parentInjector != null) {
-      parentInjector.inject(target, name);
-      return;
-    }
-
-    Injector injector = rootDependencyResolver["injector"][qualifier];
-    if (injector != null) {
-      injector.inject(target, this);
-    }
-  }
 
   FoundInjection findParrent(BuildContext context) {
     Element element = (context as BuildContext)
-        .getElementForInheritedWidgetOfExactType<InjectHost>();
+        .getElementForInheritedWidgetOfExactType<ScopeInjectHost>();
+
+    if (context is StatelessElement) {
+      ScopeInjectHost widget = element.widget as ScopeInjectHost;
+      return FoundInjection(widget.getIt(), element);
+    }
 
     element.visitAncestorElements((el) {
       element = el;
       return false;
     });
 
-    element = element.getElementForInheritedWidgetOfExactType<InjectHost>();
+    element =
+        element.getElementForInheritedWidgetOfExactType<ScopeInjectHost>();
     if (element == null) {
       return FoundInjection(null, null);
     }
 
-    InjectHost widget = element.widget as InjectHost;
+    ScopeInjectHost widget = element.widget as ScopeInjectHost;
     return FoundInjection(widget.getIt(), element);
   }
 }