+++ finject_flutter/CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG.md
 
+## (0.1.3)
+
+Breaking Changes
+
+Breaking Changes:
+
+  - Use FinjectHost class instead of InjectHost class.
+
+Features:
+
+  - Separation of scoping and injection
+
+
 ## (0.1.2)
 
 Features: